Move log screen component to Screens
